+++ src/server/app.ts
@@ -8,6 +8,7 @@ import gameRoute = require("./routes/game");
 import jsRoute = require("./routes/js");
 import sandboxConfigRoute = require("./routes/sandboxConfig");
 import testRoute = require("./routes/test");
+import { resolveEngineFilesVariable } from "./utils";
 
 interface AkashicSandbox extends express.Express {
 	gameBase?: string;
@@ -38,8 +39,8 @@ interface GameConfiguration {
 }
 
 function result2csv(results: any[]): string {
-	var csv: string = "";
-	for (var i = 0; i < results.length; i++) {
+	let csv: string = "";
+	for (let i = 0; i < results.length; i++) {
 		csv += results[i].name + "," + results[i].elapse + "\n";
 	}
 	return csv;
@@ -47,26 +48,26 @@ function result2csv(results: any[]): string {
 
 function getContentModuleEnvironment(gameJsonPath: string): ModuleEnvironment {
 	if (fs.existsSync(gameJsonPath)) {
-		var configuration: GameConfiguration = JSON.parse(fs.readFileSync(gameJsonPath, "utf8"));
+		const configuration: GameConfiguration = JSON.parse(fs.readFileSync(gameJsonPath, "utf8"));
 		return configuration.environment;
 	}
 	return null;
 }
 
 module.exports = function (options: AppOptions = {}): AkashicSandbox {
-	var appBase = path.join(__dirname, "..");
-	var gameBase = options.gameBase ? options.gameBase : process.cwd();
-	var cascadeBases = options.cascadeBases || [];
-	var jsBase = options.jsBase ? options.jsBase : path.join(appBase, "js");
-	var cssBase = options.cssBase ? options.cssBase : path.join(appBase, "css");
-	var thridpartyBase = options.thirdpartyBase ? options.thirdpartyBase : path.join(appBase, "thirdparty");
+	const appBase = path.join(__dirname, "..");
+	const gameBase = options.gameBase ? options.gameBase : process.cwd();
+	const cascadeBases = options.cascadeBases || [];
+	const jsBase = options.jsBase ? options.jsBase : path.join(appBase, "js");
+	const cssBase = options.cssBase ? options.cssBase : path.join(appBase, "css");
+	const thridpartyBase = options.thirdpartyBase ? options.thirdpartyBase : path.join(appBase, "thirdparty");
 
-	var app: AkashicSandbox = express();
-	var isDev = app.get("env") === "development";
+	const app: AkashicSandbox = express();
+	const isDev = app.get("env") === "development";
 
-	var gameJsonPath = path.join(gameBase, "game.json");
-	var environment = getContentModuleEnvironment(gameJsonPath);
-	var version = environment && environment["sandbox-runtime"] ? environment["sandbox-runtime"] : "1";
+	const gameJsonPath = path.join(gameBase, "game.json");
+	const environment = getContentModuleEnvironment(gameJsonPath);
+	const version = environment && environment["sandbox-runtime"] ? environment["sandbox-runtime"] : "1";
 
 	if (! /^(1|2|3)$/.test(version)) {
 		// sandbox-runtime の値が "1", "2", "3" 以外の場合エラーとする
@@ -136,19 +137,19 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 	app.use("/raw_game/", express.static(app.gameBase));
 	app.use("/cascade/:index", <express.RequestHandler>jsRoute);
 	app.use("/raw_cascade/:index", <express.RequestHandler>jsRoute);
-	for (var i = 0; i < app.cascadeBases.length; ++i) {
+	for (let i = 0; i < app.cascadeBases.length; ++i) {
 		app.use("/cascade/" + i + "/", express.static(app.cascadeBases[i]));
 		app.use("/raw_cascade/" + i + "/", express.static(app.cascadeBases[i]));
 	}
 
 	app.use("/configuration/", (req: express.Request, res: express.Response, _next: Function) => {
-		var prefix = req.query.raw ? "/raw_" : "/";
+		const prefix = req.query.raw ? "/raw_" : "/";
 		if (app.cascadeBases.length === 0) {
 			res.redirect(prefix + "game/game.json");
 			return;
 		}
-		var defs = [prefix + "game/game.json"];
-		for (var i = 0; i < app.cascadeBases.length; ++i)
+		const defs = [prefix + "game/game.json"];
+		for (let i = 0; i < app.cascadeBases.length; ++i)
 			defs.push(prefix + "cascade/" + i + "/game.json");
 		res.json({ definitions: defs });
 	});
@@ -158,22 +159,20 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 	});
 
 	app.use("/engine", (req: express.Request, res: express.Response, _next: Function) => {
-		var host = req.protocol + "://" + req.get("host");
+		const host = req.protocol + "://" + req.get("host");
 		res.type("application/json");
 
-		var externals = req.query.externals ? req.query.externals : ["audio", "xhr", "websocket"];
+		let externals = req.query.externals ? req.query.externals : ["audio", "xhr", "websocket"];
 
 		externals = Array.isArray(externals) ? externals : [externals] as string[];
 
 		if (typeof externals[0] !== "string" && externals[0] != null) throw new Error("Invalid externals type");
-		// json の読み込みのため require の lint エラーを抑止
-		/* eslint-disable @typescript-eslint/no-var-requires */
-		var versionsJson = require("./engineFilesVersion.json");
+		const engineFilesVariable = resolveEngineFilesVariable(version);
 		res.render("engine", {
 			host: host,
 			version: version,
 			externals: JSON.stringify(externals),
-			engineFilesVariable: versionsJson[`v${version}`].variable
+			engineFilesVariable: engineFilesVariable
 		});
 	});
 
@@ -181,22 +180,22 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 		res.redirect("/test/");
 	});
 	app.use("/test/*.js$", (req: sr.ScriptRequest, _res: express.Response, next: Function) => {
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 		req.baseDir = app.scenario.benchmarks[ssn.cntr].target;
 		next();
 	});
 	app.use("/start/", (req: express.Request, res: express.Response, _next: Function) => {
-		var scenarioJSONString = fs.readFileSync(app.settings.scenarioPath).toString();
+		const scenarioJSONString = fs.readFileSync(app.settings.scenarioPath).toString();
 		app.scenario = JSON.parse(scenarioJSONString);
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 		ssn.cntr = 0;
 		ssn.results = [];
 		res.redirect("/test/");
 	});
 	app.use("/next/", (req: express.Request, res: express.Response, _next: Function) => {
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 
-		var elapse: number = Number(req.query.elapse);
+		const elapse: number = Number(req.query.elapse);
 		ssn.results.push(
 			{
 				"name": app.scenario.benchmarks[ssn.cntr].name,
@@ -212,7 +211,7 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 		}
 	});
 	app.use("/finish/", (req: express.Request, res: express.Response, _next: Function) => {
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 		console.log("you arrived at 'finish'");
 
 		res.render("finish", {
@@ -223,7 +222,7 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 	});
 	app.use("/test", <express.RequestHandler>jsRoute);
 	app.use("/test", (req: express.Request, res: express.Response, next: express.NextFunction) => {
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 		res.locals.maxAge = app.scenario.benchmarks[ssn.cntr].maxAge;
 		res.locals.renderPerFrame = app.scenario.benchmarks[ssn.cntr].renderPerFrame;
 		res.locals.renderPerFrame = (res.locals.renderPerFrame === undefined) ? "undefined" : res.locals.renderPerFrame;
@@ -232,17 +231,17 @@ module.exports = function (options: AppOptions = {}): AkashicSandbox {
 		(<express.RequestHandler>testRoute)(req, res, next);
 	});
 	app.use("/test/", (req: express.Request, res: express.Response, next: express.NextFunction) => {
-		var ssn: ASSession = req.session;
+		const ssn: ASSession = req.session;
 		express.static(app.scenario.benchmarks[ssn.cntr].target)(req, res, next);
 	});
 
 	app.use((_req: express.Request, _res: express.Response, next: Function) => {
-		var err = new Error("Not Found");
+		const err = new Error("Not Found");
 		err.status = 404;
 		next(err);
 	});
 
-	var errorHandler: express.ErrorRequestHandler;
+	let errorHandler: express.ErrorRequestHandler;
 	if (isDev) {
 		errorHandler = (err: any, _req: express.Request, res: express.Response, _next: Function): any => {
 			res.status(err.status || 500);