+++ src/server/controller/sandboxConfig.ts
@@ -4,15 +4,15 @@ import express = require("express");
 import sr = require("../request/ScriptRequest");
 
 
-var controller: express.RequestHandler = (req: sr.ScriptRequest, res: express.Response, _next: Function) => {
-	var scriptPath = path.resolve(path.join(req.baseDir, "sandbox.config.js"));
+const controller: express.RequestHandler = (req: sr.ScriptRequest, res: express.Response, _next: Function) => {
+	const scriptPath = path.resolve(path.join(req.baseDir, "sandbox.config.js"));
 
 	if (! fs.existsSync(scriptPath)) {
 		res.contentType("text/javascript");
 		res.send(createLoadingScript({}));
 		return;
 	}
-	var sandboxConfig = {};
+	let sandboxConfig = {};
 	try {
 		sandboxConfig = require(scriptPath);
 		delete require.cache[require.resolve(scriptPath)]; // 設定ファイルの更新に追従するためアクセス毎にキャッシュを削除する
@@ -27,7 +27,7 @@ var controller: express.RequestHandler = (req: sr.ScriptRequest, res: express.Re
 module.exports = controller;
 
 function completeConfigParams(c: SandboxConfig): SandboxConfig {
-	var config = {
+	const config = {
 		autoSendEventName: c.autoSendEventName ? c.autoSendEventName : "",
 		events: c.events ? c.events : {},
 		showMenu: !!c.showMenu,